Encapsulate Hub logic one step further
Always think/design your code in library, or in go dialect: packages.

Even if that's just an toy project, having clear separation of concerns
between the main function and the rest of the code helps.

In the main function I setup various components of my applications.
Those components may come from packages, files, or even structs in the
same file, but it's better to try and keep the main function tidy,
minimal.

2 things here:

-1-
Hub has a Run function, so as a user of your library/app, I'm expecting
that calling hub.Run() is sufficient to start the hub.
Starting the hub means handling incoming WS connections, as well as
treating user interaction specific cases in those connections.

Run starts it all.

-2-

Better move the concurrent stuff at the lower level possible. Do not
expose them. This is linked to -1-. That's why I started the goroutine
inside hub.Run, so that's if Hub was a library, I'd just have to call
hub.Run, without having to deal with concurrent stuff. To sum it up:
better hide channel/goroutines/etc. from the high-level API.